chore: removes index file from Input exports
