+++ src/components/index.ts
@@ -11,7 +11,7 @@ export { Slider, type ISliderProps } from './data-entry/Slider/Slider'
 export { Cascader, type ICascaderProps } from './data-entry/Cascader/Cascader'
 export { DatePicker, type IDatePickerProps } from './data-entry/DatePicker/DatePicker'
 export { Checkbox, type ICheckboxProps } from './data-entry/Checkbox/Checkbox'
-export { Input, type IInputProps, type InputRef } from './data-entry/Input'
+export { Input, type IInputProps, type InputRef } from './data-entry/Input/Input'
 export { InputNumber, type IInputNumberProps } from './data-entry/InputNumber/InputNumber'
 export { Switch, type ISwitchProps } from './data-entry/Switch/Switch'
 export { Upload, type IUploadProps } from './data-entry/Upload/Upload'
